+++ example/lib/models/adapters.dart
@@ -3,6 +3,6 @@ import 'package:hive_local_storage/hive_local_storage.dart';
 
 part 'adapters.g.dart';
 
-@GenerateAdapters([AdapterSpec<User>()], firstTypeId: 1)
+@GenerateAdapters([AdapterSpec<User>()])
 // ignore: unused_element
 void _() {}