+++ lib/hive_local_storage.dart
@@ -5,5 +5,6 @@ library;
 export 'package:hive_ce/hive.dart';
 export 'package:hive_ce_flutter/hive_flutter.dart';
 
+export 'src/auth_token.dart';
+export 'src/encryption_helper.dart';
 export 'src/hive_local_storage.dart';
-export 'src/session.dart';