+++ lib/src/auth_token.g.dart
@@ -1,0 +1,28 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'auth_token.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_AuthToken _$AuthTokenFromJson(Map<String, dynamic> json) => _AuthToken(
+  accessToken: json['accessToken'] as String,
+  refreshToken: json['refreshToken'] as String? ?? '',
+  createdAt:
+      json['createdAt'] == null
+          ? null
+          : DateTime.parse(json['createdAt'] as String),
+  updatedAt:
+      json['updatedAt'] == null
+          ? null
+          : DateTime.parse(json['updatedAt'] as String),
+);
+
+Map<String, dynamic> _$AuthTokenToJson(_AuthToken instance) =>
+    <String, dynamic>{
+      'createdAt': instance.createdAt.toIso8601String(),
+      'updatedAt': instance.updatedAt.toIso8601String(),
+      'accessToken': instance.accessToken,
+      'refreshToken': instance.refreshToken,
+    };