+++ lib/src/hive/adapters.dart
@@ -1,7 +1,0 @@
-import 'package:hive_local_storage/hive_local_storage.dart';
-
-part 'adapters.g.dart';
-
-@GenerateAdapters([AdapterSpec<Session>()])
-// ignore: unused_element
-void _() {}