+++ lib/src/hive_local_storage.dart
@@ -1,19 +1,13 @@
 import 'dart:async';
 import 'dart:convert';
 
-// ignore: depend_on_referenced_packages
-import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:hive_ce/hive.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
-import 'package:rxdart/rxdart.dart';
 import 'package:synchronized/synchronized.dart';
 
-import '_encryption_helper.dart';
-import '_jwt_decoder.dart';
-import 'hive/hive_registrar.g.dart';
-import 'session.dart';
+import 'encryption_helper.dart';
 
 /// {@template local_storage}
 /// A wrapper class for session and cache box uses [Hive]
@@ -53,14 +47,8 @@ class LocalStorage {
 
   static final _lock = Lock();
 
-  /// session key
-  static const String _sessionKey = '__JWT_SESSION_KEY__';
-
   /// cache key
   static const String _cacheKey = '__CACHE_KEY__';
-
-  /// [Box] session box
-  static late Box<Session> _sessionBox;
 
   /// [Box] _cacheBox
   static late Box<dynamic> _cacheBox;
@@ -83,30 +71,36 @@ class LocalStorage {
   /// open the boxes
   /// returns [LocalStorage] instance
   static Future<LocalStorage> getInstance({
-    String? storageDirectory,
+    String storageDirectory = '.session',
     void Function()? registerAdapters,
-    HiveCipher? customCipher,
+    HiveCipher? encryptionCipher,
   }) async {
     await _lock.synchronized(() async {
-      final appDir = await getApplicationDocumentsDirectory();
-      if (!kIsWeb) {
-        _storagePath = appDir.path;
-        _storagePath = p.join(_storagePath!, storageDirectory);
+      try {
+        final appDir = await getApplicationDocumentsDirectory();
+        if (!kIsWeb) {
+          _storagePath = appDir.path;
+          _storagePath = p.join(_storagePath!, storageDirectory);
+        }
+        Hive.init(_storagePath);
+        registerAdapters?.call();
+        await _initBoxes(encryptionCipher);
+      } catch (_) {
+        await Hive.deleteFromDisk();
+        await Hive.close();
+        await _initBoxes(encryptionCipher);
       }
-      Hive.init(_storagePath);
-      Hive.registerAdapters();
-      registerAdapters?.call();
-      _encryptionCipher = await EncryptionHelper.hiveCipher(customCipher);
-      _sessionBox = await Hive.openBox<Session>(
-        _sessionKey,
-        encryptionCipher: _encryptionCipher,
-      );
-      _cacheBox = await Hive.openBox(
-        _cacheKey,
-        encryptionCipher: _encryptionCipher,
-      );
     });
     return LocalStorage._();
+  }
+
+  /// initialize boxes
+  static Future<void> _initBoxes(HiveCipher? customCipher) async {
+    _encryptionCipher = await EncryptionHelper.hiveCipher(customCipher);
+    _cacheBox = await Hive.openBox(
+      _cacheKey,
+      encryptionCipher: _encryptionCipher,
+    );
   }
 
   /// `openBox`
@@ -230,88 +224,6 @@ class LocalStorage {
     });
   }
 
-  /// private getter to access session
-  Session? get _session => _guard(() => _sessionBox.values.firstOrNull);
-
-  /// accessToken
-  /// getter to access accessToken
-  String? get accessToken => _session?.accessToken;
-
-  /// refreshToken
-  /// getter to access refreshToken
-  String? get refreshToken => _session?.refreshToken;
-
-  /// createdAt
-  /// getter to access createdAt
-  DateTime? get createdAt => _session?.createdAt;
-
-  /// updatedAt
-  /// getter to access updatedAt
-  DateTime? get updatedAt => _session?.updatedAt;
-
-  /// remainingTokenDuration
-  /// returns remaining time of token
-  Duration get remainingTokenDuration {
-    return _guard(() {
-      if (!hasSession) return Duration.zero;
-      return JwtDecoder.getRemainingTime(_session!.accessToken);
-    });
-  }
-
-  /// `onSessionChange`
-  /// returns stream of [bool] when data changes on box
-  Stream<bool> get onSessionChange {
-    return _guard(
-      () => _sessionBox
-          .watch()
-          .distinct()
-          .map<bool>((event) {
-            return event.value != null;
-          })
-          .startWith(hasSession),
-      () => Stream.value(false),
-    );
-  }
-
-  /// `hasSession`
-  /// checks whether `Box<Session>` is not empty or [Session] is not null
-  bool get hasSession {
-    return _guard(
-      () => _sessionBox.isNotEmpty && _session != null,
-      () => false,
-    );
-  }
-
-  /// `saveToken`
-  /// updates access token if exists or saves new one if not
-  /// updates refresh token
-  Future<void> saveToken(String token, [String refreshToken = '']) {
-    return _lockGuard(() async {
-      final session = Session(
-        accessToken: token,
-        refreshToken: refreshToken,
-        createdAt: DateTime.now(),
-      );
-      await _sessionBox.clear();
-      await _sessionBox.add(session);
-    });
-  }
-
-  /// `isTokenExpired`
-  /// checks whether token is expired or not
-  bool? get isTokenExpired {
-    return _guard(() {
-      if (!hasSession) return null;
-      return JwtDecoder.isExpired(_session!.accessToken);
-    });
-  }
-
-  /// clearSession
-  /// removes the [Session] value from [Box]
-  Future<void> clearSession() {
-    return _lockGuard(_sessionBox.clear);
-  }
-
   /// watch
   /// watch specific key for value changed
   Stream<T?> watchKey<T>({required String key, String? boxName}) {
@@ -364,7 +276,7 @@ class LocalStorage {
 
   /// `writeAndClose`
   /// write value to box and close the box
-  static Future<void> writeAndClose<T>({
+  Future<void> writeAndClose<T>({
     required String boxName,
     required String key,
     required T value,
@@ -390,7 +302,7 @@ class LocalStorage {
 
   /// `readAndClose`
   /// read value from box and close the box
-  static Future<T?> readAndClose<T>({
+  Future<T?> readAndClose<T>({
     required String key,
     required String boxName,
   }) async {
@@ -420,26 +332,20 @@ class LocalStorage {
   /// clear all values from both session box and cache box
   /// does not clears box created using `openBox()`
   Future<void> clearAll() async {
-    return _lockGuard(() {
-      Future.wait([_sessionBox.clear(), _cacheBox.clear()]);
-    });
+    return _lockGuard(_cacheBox.clear);
   }
 
   /// close all the opened box
   Future<void> closeAll() async {
     return _lockGuard(() {
-      Future.wait([_sessionBox.close(), _cacheBox.close(), Hive.close()]);
+      Future.wait([_cacheBox.close(), Hive.close()]);
     });
   }
 
   /// delete all the opened box
   Future<void> deleteAll() {
     return _lockGuard(() {
-      Future.wait([
-        _sessionBox.deleteFromDisk(),
-        _cacheBox.deleteFromDisk(),
-        Hive.deleteFromDisk(),
-      ]);
+      Future.wait([_cacheBox.deleteFromDisk(), Hive.deleteFromDisk()]);
     });
   }